requirements.txt
streamlit
azure-storage-blob
azure-ai-textanalytics
scikit-learn
pdfplumber
python-docx